Potential fix for code scanning alert no. 2: Implicit narrowing conversion in compound assignment

+++ src/main/java/org/mesutormanli/visualnovel/engine/util/RelativeLayout.java
@@ -415,7 +415,7 @@ public class RelativeLayout implements LayoutManager2, java.io.Serializable {
                 Dimension d = component.getPreferredSize();
                 spaceAvailable -= d.width;
             } else {
-                relativeTotal += constraint.doubleValue();
+                relativeTotal += constraint.floatValue();
             }
         }
